% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robma_sensitivity.R
\name{attach_robma_sensitivity}
\alias{attach_robma_sensitivity}
\title{Attach RoBMA Sensitivity Fits to a bayesma Object}
\usage{
attach_robma_sensitivity(model, robma_sensitivity)
}
\arguments{
\item{model}{A \code{bayesma} object.}

\item{robma_sensitivity}{A \code{bayesma_robma_sensitivity} object created by
\code{robma_sensitivity()}.}
}
\value{
The modified \code{bayesma} object with \verb{$robma_sensitivity} attached.
}
\description{
Attaches pre-computed RoBMA sensitivity fits to a bayesma object.
These fits will be used by \code{sensitivity_plot()} when \code{incl_robma = TRUE}.
}
