% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesma_mv.R
\name{bayesma_mv_extract}
\alias{bayesma_mv_extract}
\title{Extract tidy effect components from a bivariate meta-analysis fit}
\usage{
bayesma_mv_extract(fit, spec)
}
\arguments{
\item{fit}{A \code{bayesma_mv_fit} object.}

\item{spec}{A \code{bayesma_mv_spec} object.}
}
\value{
An object of class \code{"bayesma_mv_effects"}.
}
\description{
Extract tidy effect components from a bivariate meta-analysis fit
}
