% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesma_mv.R
\name{bayesma_mv_fit}
\alias{bayesma_mv_fit}
\title{Compile and sample a bivariate meta-analysis model}
\usage{
bayesma_mv_fit(
  spec,
  code = bayesma_mv_stan_code(spec),
  stan_data = bayesma_mv_stan_data(spec),
  chains = 4,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt_delta = 0.95,
  seed = 1234,
  ...
)
}
\arguments{
\item{spec}{A \code{bayesma_mv_spec} object.}

\item{code}{A \code{bayesma_mv_stan_code} object. Defaults to
\code{bayesma_mv_stan_code(spec)}.}

\item{stan_data}{A Stan data list. Defaults to \code{bayesma_mv_stan_data(spec)}.}

\item{chains, iter_warmup, iter_sampling, adapt_delta, seed}{MCMC settings.}

\item{...}{Passed to \code{cmdstanr::CmdStanModel$sample()}.}
}
\value{
An object of class \code{"bayesma_mv_fit"}.
}
\description{
Compile and sample a bivariate meta-analysis model
}
