% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesma_mv.R
\name{bayesma_mv_output}
\alias{bayesma_mv_output}
\title{Assemble a \code{bayesma_mv} object from pipeline outputs}
\usage{
bayesma_mv_output(spec, fit, effects)
}
\arguments{
\item{spec}{A \code{bayesma_mv_spec} object.}

\item{fit}{A \code{bayesma_mv_fit} object.}

\item{effects}{A \code{bayesma_mv_effects} object.}
}
\value{
A list of class \code{c("bayesma_mv", "bayesma")}.
}
\description{
Assemble a \code{bayesma_mv} object from pipeline outputs
}
