% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesma_pipeline.R
\name{bayesma_output}
\alias{bayesma_output}
\title{Assemble a \code{bayesma} object from pipeline outputs}
\usage{
bayesma_output(spec, fit, effects)
}
\arguments{
\item{spec}{A \code{bayesma_spec} object.}

\item{fit}{A \code{bayesma_fit} object.}

\item{effects}{A \code{bayesma_effects} object (from \code{\link[=bayesma_extract]{bayesma_extract()}}).}
}
\value{
A list of class \code{"bayesma"}.
}
\description{
Combines the outputs of the earlier stages into the standard \code{bayesma}
return object consumed by downstream tools (\code{forest()}, \code{diagnostics()},
etc.).
}
