% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger.R
\name{egger_extract}
\alias{egger_extract}
\title{Extract tidy summaries from an Egger's test fit}
\usage{
egger_extract(fit, spec)
}
\arguments{
\item{fit}{An \code{egger_fit} object.}

\item{spec}{An \code{egger_spec} object.}
}
\value{
An object of class \code{"egger_effects"}.
}
\description{
Extract tidy summaries from an Egger's test fit
}
