% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger.R
\name{egger_fit}
\alias{egger_fit}
\title{Compile and sample an Egger's test model}
\usage{
egger_fit(
  spec,
  code = egger_stan_code(spec),
  stan_data = egger_stan_data(spec),
  chains = 4,
  iter_warmup = 2000,
  iter_sampling = 4000,
  adapt_delta = 0.95,
  seed = 1234,
  ...
)
}
\arguments{
\item{spec}{An \code{egger_spec} object.}

\item{code}{An \code{egger_stan_code} object. Defaults to
\code{egger_stan_code(spec)}.}

\item{stan_data}{A Stan data list. Defaults to \code{egger_stan_data(spec)}.}

\item{chains, iter_warmup, iter_sampling, adapt_delta, seed}{MCMC settings.}

\item{...}{Passed to \code{cmdstanr::CmdStanModel$sample()}.}
}
\value{
An object of class \code{"egger_fit"}.
}
\description{
Compile and sample an Egger's test model
}
