% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger.R
\name{egger_output}
\alias{egger_output}
\title{Assemble a \code{bayesma_egger} object from pipeline outputs}
\usage{
egger_output(spec, fit, effects)
}
\arguments{
\item{spec}{An \code{egger_spec} object.}

\item{fit}{An \code{egger_fit} object.}

\item{effects}{An \code{egger_effects} object.}
}
\value{
A list of class \code{"bayesma_egger"}.
}
\description{
Assemble a \code{bayesma_egger} object from pipeline outputs
}
