% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger.R
\name{egger_stan_data}
\alias{egger_stan_data}
\title{Build the Stan data list for an Egger's test specification}
\usage{
egger_stan_data(spec)
}
\arguments{
\item{spec}{An \code{egger_spec} object.}
}
\value{
A named list suitable for \code{cmdstanr::CmdStanModel$sample(data = ...)}.
}
\description{
Build the Stan data list for an Egger's test specification
}
