% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_sr_plot.R
\name{inspect_summary_plot}
\alias{inspect_summary_plot}
\title{Create INSPECT-SR Summary Bar Plot}
\usage{
inspect_summary_plot(data, font = NULL, title = NULL)
}
\arguments{
\item{data}{An \code{inspect_sr} object or data frame.}

\item{font, title}{As in \code{\link{inspect_plot}}.}
}
\value{
A ggplot2 object.
}
\description{
Create INSPECT-SR Summary Bar Plot
}
