% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_reg.R
\name{meta_reg_extract}
\alias{meta_reg_extract}
\title{Extract tidy effect components from a meta-regression fit}
\usage{
meta_reg_extract(fit, spec)
}
\arguments{
\item{fit}{A \code{meta_reg_fit} object.}

\item{spec}{A \code{meta_reg_spec} object.}
}
\value{
An object of class \code{"meta_reg_effects"}.
}
\description{
Extract tidy effect components from a meta-regression fit
}
