% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_reg.R
\name{meta_reg_fit}
\alias{meta_reg_fit}
\title{Compile and sample a meta-regression model}
\usage{
meta_reg_fit(
  spec,
  code = meta_reg_stan_code(spec),
  stan_data = meta_reg_stan_data(spec),
  chains = 4,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt_delta = 0.95,
  seed = 1234,
  ...
)
}
\arguments{
\item{spec}{A \code{meta_reg_spec} object.}

\item{code}{A \code{meta_reg_stan_code} object. Defaults to
\code{meta_reg_stan_code(spec)}.}

\item{stan_data}{A Stan data list. Defaults to \code{meta_reg_stan_data(spec)}.}

\item{chains, iter_warmup, iter_sampling, adapt_delta, seed}{MCMC settings.}

\item{...}{Passed to \code{cmdstanr::CmdStanModel$sample()}.}
}
\value{
An object of class \code{"meta_reg_fit"}.
}
\description{
Compile and sample a meta-regression model
}
