% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_reg.R
\name{meta_reg_output}
\alias{meta_reg_output}
\title{Assemble a \code{bayesma_metareg} object from pipeline outputs}
\usage{
meta_reg_output(spec, fit, effects)
}
\arguments{
\item{spec}{A \code{meta_reg_spec} object.}

\item{fit}{A \code{meta_reg_fit} object.}

\item{effects}{A \code{meta_reg_effects} object.}
}
\value{
A list of class \code{c("bayesma_metareg", "bayesma")}.
}
\description{
Assemble a \code{bayesma_metareg} object from pipeline outputs
}
