% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesma_helpers.R
\name{print.bayesma}
\alias{print.bayesma}
\alias{print.bayesma_summary}
\title{Print a summary for a fitted model represented by a \code{bayesma} object}
\usage{
\method{print}{bayesma}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{bayesma}.}

\item{digits}{The number of significant digits for printing out the summary;
defaults to 2. Bulk_ESS and Tail_ESS are always rounded to integers.}

\item{...}{Additional arguments passed to \code{\link{summary.bayesma}}.}
}
\description{
Print a summary for a fitted model represented by a \code{bayesma} object
}
\seealso{
\code{\link{summary.bayesma}}
}
\keyword{internal}
