% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_plot.R
\name{render_sensitivity_patchwork}
\alias{render_sensitivity_patchwork}
\title{Render the stored components into a patchwork object}
\usage{
render_sensitivity_patchwork(x)
}
\arguments{
\item{x}{A \code{bayesma_sensitivity_plot} object.}
}
\value{
A patchwork object.
}
\description{
Render the stored components into a patchwork object
}
