% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robma_pipeline.R
\name{robma_extract}
\alias{robma_extract}
\title{Extract tidy effect components from a RoBMA fit}
\usage{
robma_extract(fit, spec)
}
\arguments{
\item{fit}{A \code{bayesma_robma_fit}.}

\item{spec}{A \code{bayesma_robma_spec}.}
}
\value{
An object of class \code{"bayesma_robma_effects"}.
}
\description{
Extract tidy effect components from a RoBMA fit
}
