% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robma_pipeline.R
\name{robma_output}
\alias{robma_output}
\title{Assemble the final RoBMA output object}
\usage{
robma_output(spec, fit, effects)
}
\arguments{
\item{spec}{A \code{bayesma_robma_spec}.}

\item{fit}{A \code{bayesma_robma_fit}.}

\item{effects}{A \code{bayesma_robma_effects}.}
}
\value{
An object of class \code{c("bayesma_robma", "bayesma")}.
}
\description{
Assemble the final RoBMA output object
}
