% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesma_helpers.R
\name{stan_code}
\alias{stan_code}
\title{Print Stan code from a fitted model}
\usage{
stan_code(fit)
}
\arguments{
\item{fit}{A \code{CmdStanFit} object.}
}
\value{
The Stan code string (invisibly).
}
\description{
Print Stan code from a fitted model
}
