% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egger.R
\name{summary.bayesma_egger}
\alias{summary.bayesma_egger}
\title{Summarise a fitted \code{bayesma_egger} model}
\usage{
\method{summary}{bayesma_egger}(object, ...)
}
\arguments{
\item{object}{A \code{bayesma_egger} object.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{egger_summary}.
}
\description{
Summarise a fitted \code{bayesma_egger} model
}
\keyword{internal}
